new arthmetic opretor and unary ops

+++ interpreter/interpreter.h
@@ -3,7 +3,7 @@
 
 #include "..\parser\parser.h"
 
-// Function to interpret the AST and evaluate the expression
-int interpret(ASTNode* node);
+// Function to interpret an AST and return the result
+double interpret(ASTNode* node);
 
 #endif /* INTERPRETER_H */
